refactor: rename tools and move ip.go and port.go ot shared
