+++ server/cmd/auth/pkg/md5.go
@@ -1,4 +1,4 @@
-package tool
+package pkg
 
 import (
 	"crypto/md5"
