+++ server/cmd/blob/initialize/db.go
@@ -2,10 +2,10 @@ package initialize
 
 import (
 	"fmt"
-	"github.com/CyanAsterisk/FreeCar/server/shared/consts"
 	"time"
 
 	"github.com/CyanAsterisk/FreeCar/server/cmd/blob/global"
+	"github.com/CyanAsterisk/FreeCar/server/shared/consts"
 	"github.com/cloudwego/kitex/pkg/klog"
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
