+++ server/cmd/car/global/global.go
@@ -2,7 +2,7 @@ package global
 
 import (
 	"github.com/CyanAsterisk/FreeCar/server/cmd/car/config"
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/tool/mq"
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/pkg/mq"
 	"github.com/CyanAsterisk/FreeCar/server/shared/kitex_gen/car/carservice"
 	"github.com/CyanAsterisk/FreeCar/server/shared/kitex_gen/trip/tripservice"
 	"go.mongodb.org/mongo-driver/mongo"
