+++ server/cmd/car/initialize/nacos.go
@@ -1,9 +1,10 @@
 package initialize
 
 import (
-	"github.com/CyanAsterisk/FreeCar/server/shared/consts"
 	"net"
 	"strconv"
+
+	"github.com/CyanAsterisk/FreeCar/server/shared/consts"
 
 	"github.com/CyanAsterisk/FreeCar/server/cmd/car/global"
 	"github.com/bwmarrin/snowflake"
@@ -74,6 +75,9 @@ func InitNacos(Port int) (registry.Registry, *registry.Info) {
 			ServerConfigs: sc,
 		},
 	)
+	if err != nil {
+		klog.Fatalf("new naming client failed: %s", err.Error())
+	}
 
 	r := nacos.NewNacosRegistry(registryClient, nacos.WithGroup(consts.CarGroup))
