+++ server/cmd/car/main.go
@@ -5,15 +5,14 @@ import (
 	"net"
 	"strconv"
 
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/global"
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/initialize"
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/pkg/sim"
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/pkg/trip"
+	"github.com/CyanAsterisk/FreeCar/server/cmd/car/pkg/ws"
 	"github.com/CyanAsterisk/FreeCar/server/shared/consts"
 	"github.com/CyanAsterisk/FreeCar/server/shared/kitex_gen/car/carservice"
 	"github.com/CyanAsterisk/FreeCar/server/shared/middleware"
-
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/global"
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/initialize"
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/tool/sim"
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/tool/trip"
-	"github.com/CyanAsterisk/FreeCar/server/cmd/car/tool/ws"
 	hzserver "github.com/cloudwego/hertz/pkg/app/server"
 	"github.com/cloudwego/kitex/pkg/klog"
 	"github.com/cloudwego/kitex/pkg/limit"
