+++ server/cmd/trip/pkg/poi/poi.go
